+++ packages/permissions-provider-snap/src/orchestrators/orchestrators.ts
@@ -1,14 +1,22 @@
 /* eslint-disable @typescript-eslint/no-throw-literal */
-import type { Permission } from '@metamask/7715-permissions-shared/types';
+import {
+  createRootDelegation,
+  encodeDelegation,
+} from '@metamask-private/delegator-core-viem';
 import {
   zNativeTokenStreamPermission,
-  type PermissionResponse,
+  type Permission,
 } from '@metamask/7715-permissions-shared/types';
 import { extractZodError } from '@metamask/7715-permissions-shared/utils';
-import { InvalidParamsError, type SnapsProvider } from '@metamask/snaps-sdk';
+import { InvalidParamsError } from '@metamask/snaps-sdk';
+import type { Hex } from 'viem';
+import { fromHex } from 'viem';
 
+import { type MockAccountController } from '../accountController.mock';
+import { type PermissionConfirmationRenderHandler } from '../ui';
 import type {
   OrchestrateMeta,
+  OrchestrateResult,
   Orchestrator,
   PermissionTypeMapping,
   SupportedPermissionTypes,
@@ -19,6 +27,26 @@ import type {
  */
 const zodObjectMapper = {
   'native-token-stream': zNativeTokenStreamPermission,
+};
+
+/**
+ * Prepare the account details for the permission picker UI.
+ * @param accountController - An account controller instance.
+ * @param chainId - The chain ID.
+ * @returns The account address, balance.
+ */
+export const prepareAccountDetails = async (
+  accountController: MockAccountController,
+  chainId: number,
+): Promise<[Hex, Hex]> => {
+  return await Promise.all([
+    accountController.getAccountAddress({
+      chainId,
+    }),
+    accountController.getAccountBalance({
+      chainId,
+    }),
+  ]);
 };
 
 /**
@@ -69,18 +97,19 @@ const validatePermissionData = <
 };
 
 /**
- * Factory function for create a permission orchestrator for a specific permission type.
+ * Factory function to create a permission orchestrator for a specific permission type.
  *
- * @param _snapsProvider - A snaps provider instance.
- * @param _accountController - An account controller instance.
+ * @param accountController - An account controller instance.
+ * @param permissionConfirmationRenderHandler - The permission confirmation render handler.
  * @param permissionType - The permission type.
- * @returns A permission orchestrator for the native-token-stream permission type.
+ * @returns A permission orchestrator for the specific permission type.
  */
 export const createPermissionOrchestrator = <
   TPermissionType extends SupportedPermissionTypes,
 >(
-  _snapsProvider: SnapsProvider,
-  _accountController: unknown,
+  // TODO: Remove mock accountController: https://app.zenhub.com/workspaces/readable-permissions-67982ce51eb4360029b2c1a1/issues/gh/metamask/delegator-readable-permissions/54
+  accountController: MockAccountController,
+  permissionConfirmationRenderHandler: PermissionConfirmationRenderHandler,
   permissionType: TPermissionType,
 ): Orchestrator<TPermissionType> => {
   return {
@@ -93,12 +122,84 @@ export const createPermissionOrchestrator = <
 
       return validatedPermission;
     },
-    orchestrate: async (_orchestrateMeta: OrchestrateMeta<TPermissionType>) => {
-      // TODO: Implement Specific permission orchestrator: https://app.zenhub.com/workspaces/readable-permissions-67982ce51eb4360029b2c1a1/issues/gh/metamask/delegator-readable-permissions/42
+    orchestrate: async (orchestrateMeta: OrchestrateMeta<TPermissionType>) => {
+      const { chainId, sessionAccount, origin, expiry, permission } =
+        orchestrateMeta;
+      const chainIdNum = fromHex(chainId, 'number');
+
+      // Get the user account details
+      const [account, balance] = await prepareAccountDetails(
+        accountController,
+        fromHex(chainId, 'number'),
+      );
+
+      // Prepare specific context and UI
+      const [context, permissionConfirmationPage] =
+        permissionConfirmationRenderHandler.getPermissionConfirmationPage(
+          {
+            permission,
+            account,
+            siteOrigin: origin,
+            balance,
+            expiry,
+            chainId: chainIdNum,
+          },
+          permissionType,
+        );
+
+      // Wait for the successful permission confirmation reponse from the user
+      const { attenuatedPermission, attenuatedExpiry, isConfirmed } =
+        await permissionConfirmationRenderHandler.getConfirmedAttenuatedPermission(
+          context,
+          permissionConfirmationPage,
+          permissionType,
+        );
+
+      if (!isConfirmed) {
+        return {
+          success: false,
+          reason: 'User rejected the permissions request',
+        };
+      }
+
+      // TODO: Use the delegation builder to attach the correct caveats specific to the permission type
+      // https://app.zenhub.com/workspaces/readable-permissions-67982ce51eb4360029b2c1a1/issues/gh/metamask/delegator-readable-permissions/41
+      const delegation = createRootDelegation(sessionAccount, account, []);
+
+      // Sign the delegation and encode it to create the permissioncContext
+      const signedDelegation = await accountController.signDelegation({
+        chainId: chainIdNum,
+        delegation,
+      });
+      const permissionContext = encodeDelegation([signedDelegation]);
+
+      const accountMeta = await accountController.getAccountMetadata({
+        chainId: chainIdNum,
+      });
+
       return {
         success: true,
-        response: {} as PermissionResponse,
-      };
+        response: {
+          chainId,
+          account,
+          expiry: attenuatedExpiry,
+          signer: {
+            type: 'account',
+            data: {
+              address: sessionAccount,
+            },
+          },
+          permission: attenuatedPermission,
+          context: permissionContext,
+          accountMeta:
+            accountMeta.factory && accountMeta.factoryData
+              ? [accountMeta]
+              : undefined,
+          signerMeta: {
+            delegationManager: '0x000000_delegation_manager', // TODO: Update to use actual values instead of mock values
+          },
+        },
+      } as OrchestrateResult;
     },
   };
 };