+++ packages/permissions-provider-snap/src/ui/components/spending-cap-details.tsx
@@ -1,14 +1,7 @@
-import type { Permission } from '@metamask/7715-permissions-shared/types';
-import type { JSXElement } from '@metamask/snaps-sdk/jsx';
+import type { SnapComponent } from '@metamask/snaps-sdk/jsx';
 import { Text, Section, Box, Icon, Tooltip } from '@metamask/snaps-sdk/jsx';
 
-type SpendingCapDetails = {
-  permission: Permission;
-};
-
-export const SpendingCapDetails: (params: SpendingCapDetails) => JSXElement = ({
-  permission: _,
-}) => {
+export const SpendingCapDetails: SnapComponent = () => {
   return (
     <Section>
       <Box direction="horizontal" alignment="space-between">