+++ packages/permissions-provider-snap/src/ui/handler/index.ts
@@ -1,0 +1,2 @@
+export * from './renderHandler';
+export * from './userInputHandler';